rockset/requirements.txt
decodablepy
paho-mqtt
kafka-python==2.0.2
python-dotenv==0.20.0
boto3
